requirements.txt
Flask==1.0.2
Flask-session ==0.3.0
bcrypt ==3.1.7
google-cloud-datastore